add starting game

+++ src/main/java/ru/spaceinvasion/mechanic/game/GamePart.kt
@@ -3,7 +3,7 @@ package ru.spaceinvasion.mechanic.game
 import ru.spaceinvasion.mechanic.game.messages.GameMessage
 
 abstract class GamePart(protected var mediator: GamePartMediator,
-                        val gamePartId: Int) {
+                        val gamePartId: Long) {
 
     abstract fun notify(message: GameMessage)
 }
